ozon_parser/parser/serializers.py
from django.contrib.auth.models import User, Group
from .models import Product
from rest_framework import serializers


class UserSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = User
        fields = ["url", "username", "email", "groups"]


class GroupSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Group
        fields = ["url", "name"]


class ProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = Product
        fields = "__all__"


class NumberOfProductsSerializer(serializers.Serializer):
    num_of_products = serializers.IntegerField(max_value=50, min_value=0)
